d3d12: Fix descriptor heaps on x86

+++ src/gallium/drivers/d3d12/d3d12_descriptor_pool.cpp
@@ -48,8 +48,8 @@ struct d3d12_descriptor_heap {
    ID3D12Device *dev;
    ID3D12DescriptorHeap *heap;
    uint32_t desc_size;
-   size_t cpu_base;
-   size_t gpu_base;
+   uint64_t cpu_base;
+   uint64_t gpu_base;
    uint32_t size;
    uint32_t next;
    util_dynarray free_list;
